backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==9.3.0
django-filter==23.3